+++ personalization-sdk/src/main/kotlin/com/personalization/di/SdkModule.kt
@@ -8,9 +8,9 @@ import com.personalization.sdk.domain.usecases.preferences.GetPreferencesValueUs
 import com.personalization.sdk.domain.usecases.preferences.SavePreferencesValueUseCase
 import com.personalization.sdk.domain.usecases.recommendation.GetRecommendedByUseCase
 import com.personalization.sdk.domain.usecases.recommendation.SetRecommendedByUseCase
-import com.personalization.features.recommendation.RecommendationManagerImpl
-import com.personalization.features.search.SearchManagerImpl
-import com.personalization.features.track_event.TrackEventManagerImpl
+import com.personalization.features.recommendation.impl.RecommendationManagerImpl
+import com.personalization.features.search.impl.SearchManagerImpl
+import com.personalization.features.track_event.impl.TrackEventManagerImpl
 import com.personalization.sdk.domain.usecases.network.ExecuteQueueTasksUseCase
 import com.personalization.sdk.domain.usecases.network.SendNetworkMethodUseCase
 import com.personalization.sdk.domain.usecases.userSettings.GetUserSettingsValueUseCase
